adopted more tests
