+++ packages/data-sql/src/query-converter/fields/fields.test.ts
@@ -2,8 +2,7 @@ import { expect, test, vi, afterAll, beforeEach } from 'vitest';
 import { convertFieldNodes, type FieldConversionResult } from './fields.js';
 import { parameterIndexGenerator } from '../param-index-generator.js';
 import type { AbstractQueryFieldNode } from '@directus/data';
-import { randomIdentifier, randomInteger } from '@directus/random';
-import type { AbstractSqlQuery } from '../../index.js';
+import { randomIdentifier } from '@directus/random';
 
 afterAll(() => {
 	vi.restoreAllMocks();
@@ -60,21 +59,19 @@ test('primitives only', () => {
 			joins: [],
 		},
 		parameters: [],
-		aliasMapping: new Map([
-			[`${randomPrimitiveField1}_RANDOM`, [randomPrimitiveFieldAlias1]],
-			[`${randomPrimitiveField2}_RANDOM`, [randomPrimitiveFieldAlias2]],
-		]),
-		nestedManys: [],
+		aliasMapping: [
+			{ type: 'root', alias: randomPrimitiveFieldAlias1, column: `${randomPrimitiveField1}_RANDOM` },
+			{ type: 'root', alias: randomPrimitiveFieldAlias2, column: `${randomPrimitiveField2}_RANDOM` },
+		],
+		subQueries: [],
 	};
 
 	const idGen = parameterIndexGenerator();
 	const result = convertFieldNodes(randomCollection, fields, idGen);
-	expect(result.clauses).toMatchObject(expected.clauses);
-	expect(result.parameters).toMatchObject(expected.parameters);
-	expect(result.aliasMapping).toMatchObject(expected.aliasMapping);
+	expect(result).toStrictEqual(expected);
 });
 
-test('primitive and function', () => {
+test('primitive, fn', () => {
 	const fields: AbstractQueryFieldNode[] = [
 		{
 			type: 'primitive',
@@ -115,18 +112,16 @@ test('primitive and function', () => {
 			joins: [],
 		},
 		parameters: [],
-		aliasMapping: new Map([
-			[`${randomPrimitiveField1}_RANDOM`, [randomPrimitiveFieldAlias1]],
-			[`month_${randomPrimitiveField2}_RANDOM`, [randomPrimitiveFieldAlias2]],
-		]),
-		nestedManys: [],
+		aliasMapping: [
+			{ type: 'root', alias: randomPrimitiveFieldAlias1, column: `${randomPrimitiveField1}_RANDOM` },
+			{ type: 'root', alias: randomPrimitiveFieldAlias2, column: `month_${randomPrimitiveField2}_RANDOM` },
+		],
+		subQueries: [],
 	};
 
 	const idGen = parameterIndexGenerator();
 	const result = convertFieldNodes(randomCollection, fields, idGen);
-	expect(result.clauses).toMatchObject(expected.clauses);
-	expect(result.parameters).toMatchObject(expected.parameters);
-	expect(result.aliasMapping).toMatchObject(expected.aliasMapping);
+	expect(result).toStrictEqual(expected);
 });
 
 test('primitive, fn, m2o', () => {
@@ -233,18 +228,28 @@ test('primitive, fn, m2o', () => {
 			],
 		},
 		parameters: [],
-		aliasMapping: new Map([
-			[`${randomPrimitiveField1}_RANDOM`, [randomPrimitiveFieldAlias1]],
-			[`${randomJoinNodeField}_RANDOM`, [randomNestedAlias, randomJoinNodeFieldAlias]],
-			[`month_${randomPrimitiveFieldFn}_RANDOM`, [randomPrimitiveFieldFnAlias]],
-		]),
-		nestedManys: [],
+		aliasMapping: [
+			{
+				type: 'root',
+				alias: randomPrimitiveFieldAlias1,
+				column: `${randomPrimitiveField1}_RANDOM`,
+			},
+			{
+				type: 'nested',
+				alias: randomNestedAlias,
+				children: [{ type: 'root', alias: randomJoinNodeFieldAlias, column: `${randomJoinNodeField}_RANDOM` }],
+			},
+			{
+				type: 'root',
+				alias: randomPrimitiveFieldFnAlias,
+				column: `month_${randomPrimitiveFieldFn}_RANDOM`,
+			},
+		],
+		subQueries: [],
 	};
 
 	const result = convertFieldNodes(randomCollection, fields, parameterIndexGenerator());
-	expect(result.clauses).toMatchObject(expected.clauses);
-	expect(result.parameters).toMatchObject(expected.parameters);
-	expect(result.aliasMapping).toMatchObject(expected.aliasMapping);
+	expect(result).toMatchObject(expected);
 });
 
 test('primitive, o2m', () => {
@@ -297,218 +302,31 @@ test('primitive, o2m', () => {
 					column: randomPrimitiveField1,
 					as: `${randomPrimitiveField1}_RANDOM`,
 				},
-			],
-			joins: [],
-		},
-		parameters: [],
-		aliasMapping: new Map([[`${randomPrimitiveField1}_RANDOM`, [randomPrimitiveFieldAlias1]]]),
-		nestedManys: [
-			{
-				queryGenerator: expect.any(Function),
-				localJoinFields: [randomJoinCurrentField],
-				foreignJoinFields: [randomExternalField],
-				alias: randomNestedAlias,
-			},
-		],
-	};
-
-	const result = convertFieldNodes(randomCollection, fields, parameterIndexGenerator());
-	expect(result).toStrictEqual(expected);
-});
-
-test('primitive and o2m field with nested modifiers', () => {
-	const randomJoinCurrentField = randomIdentifier();
-	const randomExternalCollection = randomIdentifier();
-	const randomExternalStore = randomIdentifier();
-	const randomExternalField = randomIdentifier();
-	const randomJoinNodeField = randomIdentifier();
-	const randomJoinNodeFieldAlias = randomIdentifier();
-	const randomNestedAlias = randomIdentifier();
-	const nestedJoinCurrentField = randomIdentifier();
-	const nestedExternalStore = randomIdentifier();
-	const nestedExternalCollection = randomIdentifier();
-	const nestedForeignIdFields = randomIdentifier();
-	const randomLimit = randomInteger(1, 100);
-	const compareValue = randomIdentifier();
-	const nestedTargetField = randomIdentifier();
-
-	const fields: AbstractQueryFieldNode[] = [
-		{
-			type: 'primitive',
-			field: randomPrimitiveField1,
-			alias: randomPrimitiveFieldAlias1,
-		},
-		{
-			type: 'nested-single-many',
-			fields: [
-				{
-					type: 'primitive',
-					field: randomJoinNodeField,
-					alias: randomJoinNodeFieldAlias,
-				},
-			],
-			alias: randomNestedAlias,
-			nesting: {
-				type: 'relational-many',
-				local: {
-					fields: [randomJoinCurrentField],
-				},
-				foreign: {
-					store: randomExternalStore,
-					collection: randomExternalCollection,
-					fields: [randomExternalField],
-				},
-			},
-			modifiers: {
-				filter: {
-					type: 'condition',
-					condition: {
-						type: 'condition-string',
-						operation: 'starts_with',
-						target: {
-							type: 'nested-one-target',
-							field: {
-								type: 'primitive',
-								field: nestedTargetField,
-							},
-							nesting: {
-								type: 'relational-many',
-
-								local: {
-									fields: [nestedJoinCurrentField],
-								},
-								foreign: {
-									store: nestedExternalStore,
-									collection: nestedExternalCollection,
-									fields: [nestedForeignIdFields],
-								},
-							},
-						},
-						compareTo: compareValue,
-					},
-				},
-				limit: {
-					type: 'limit',
-					value: randomLimit,
-				},
-			},
-		},
-	];
-
-	const expected: FieldConversionResult = {
-		clauses: {
-			select: [
 				{
 					type: 'primitive',
 					table: randomCollection,
-					column: randomPrimitiveField1,
-					as: `${randomPrimitiveField1}_RANDOM`,
+					column: randomJoinCurrentField,
+					as: `${randomJoinCurrentField}_RANDOM`,
 				},
 			],
 			joins: [],
 		},
 		parameters: [],
-		aliasMapping: new Map([[`${randomPrimitiveField1}_RANDOM`, [randomPrimitiveFieldAlias1]]]),
-		nestedManys: [
+		aliasMapping: [
 			{
-				queryGenerator: expect.any(Function),
-				localJoinFields: [randomJoinCurrentField],
-				foreignJoinFields: [randomExternalField],
+				type: 'root',
+				alias: randomPrimitiveFieldAlias1,
+				column: `${randomPrimitiveField1}_RANDOM`,
+			},
+			{
+				type: 'sub',
 				alias: randomNestedAlias,
+				index: 0,
 			},
 		],
-	};
-
-	const randomPkValue = randomIdentifier();
-
-	const expectedGeneratedQuery: AbstractSqlQuery = {
-		clauses: {
-			select: [
-				{
-					type: 'primitive',
-					table: randomExternalCollection,
-					column: randomJoinNodeField,
-					as: `${randomJoinNodeField}_RANDOM`,
-				},
-			],
-			from: randomExternalCollection,
-			joins: [
-				{
-					type: 'join',
-					table: nestedExternalCollection,
-					as: `${nestedExternalCollection}_RANDOM`,
-					on: {
-						type: 'condition',
-						condition: {
-							type: 'condition-field',
-							compareTo: {
-								type: 'primitive',
-								column: nestedForeignIdFields,
-								table: `${nestedExternalCollection}_RANDOM`,
-							},
-							operation: 'eq',
-							target: {
-								type: 'primitive',
-								column: nestedJoinCurrentField,
-								table: randomExternalCollection,
-							},
-						},
-						negate: false,
-					},
-				},
-			],
-			where: {
-				type: 'logical',
-				operator: 'and',
-				negate: false,
-				childNodes: [
-					{
-						type: 'condition',
-						condition: {
-							type: 'condition-string',
-							operation: 'starts_with',
-							target: {
-								type: 'primitive',
-								table: `${nestedExternalCollection}_RANDOM`,
-								column: nestedTargetField,
-							},
-							compareTo: {
-								type: 'value',
-								parameterIndex: 0,
-							},
-						},
-						negate: false,
-					},
-					{
-						type: 'condition',
-						condition: {
-							type: 'condition-string',
-							operation: 'eq',
-							target: {
-								type: 'primitive',
-								table: randomExternalCollection,
-								column: randomExternalField,
-							},
-							compareTo: {
-								type: 'value',
-								parameterIndex: 2,
-							},
-						},
-						negate: false,
-					},
-				],
-			},
-			limit: {
-				type: 'value',
-				parameterIndex: 1,
-			},
-		},
-		parameters: [compareValue, randomLimit, randomPkValue],
-		aliasMapping: new Map([[`${randomJoinNodeField}_RANDOM`, [randomJoinNodeFieldAlias]]]),
-		nestedManys: [],
+		subQueries: [expect.any(Function)],
 	};
 
 	const result = convertFieldNodes(randomCollection, fields, parameterIndexGenerator());
 	expect(result).toStrictEqual(expected);
-	expect(result.nestedManys[0]?.queryGenerator([randomPkValue])).toStrictEqual(expectedGeneratedQuery);
 });
